rename: BaseEntity 폴더 위치 수정
