+++ src/main/java/com/example/log4u/domain/diary/entity/Diary.java
@@ -1,6 +1,6 @@
 package com.example.log4u.domain.diary.entity;
 
-import com.example.log4u.domain.common.entity.BaseEntity;
+import com.example.log4u.common.entity.BaseEntity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
